gcal-mcp-server/pyproject.toml
[project]
name = "gcal-mcp-server"
version = "0.1.0"
dependencies = [
    "google-api-python-client>=2.166.0",
    "google-auth>=2.38.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.1",
    "python-dotenv>=1.1.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["gcal_mcp_server", "gcal_mcp_server.client"]
